Replace custom Maybe type with thor-option package; Add debug classes.
This commit removes the custom Maybe type and its classes (Maybe and Option), and substitutes with the external thor-option package which performs the same functionality. In addition to that, debug classes (RetryResolver and ResultState) were introduced to improve error handling procedures. All these changes are also reflected in the updates to the composer.json and composer.lock files.